Word Lock Brute Force Tool
I used this to brute force the Master Lock Word Lock that we found
in a drawer at the house. It was a fun little programming challenge
to figure out how to check which combinations were English words,
and how common those words are in English. It ended up being the 22nd
try when I opened the lock.

It could have been done easier and better other ways, but this was
definitely the most fun.